+++ src/models/lstm/model2.rs
@@ -72,8 +72,8 @@ pub struct LstmModel2<const H: usize, const H1_IH: usize, const H2_IH: usize, co
     pub layer1: LstmLayer<1, H, H1_IH, H_H4>,
     /// Model layer 2.
     pub layer2: LstmLayer<H, H, H2_IH, H_H4>,
-    /// Output head weights (quantized).
-    pub head_weights: [u16; H],
+    /// Output head weights.
+    pub head_weights: [f32; H],
     /// Output head weights in full f32 precision (mixed-precision selective).
     pub head_weights_f32: [f32; H],
     /// Output head bias.
@@ -94,7 +94,7 @@ impl<const H: usize, const H1_IH: usize, const H2_IH: usize, const H_H4: usize>
         Self {
             layer1: LstmLayer::new(),
             layer2: LstmLayer::new(),
-            head_weights: [0u16; H],
+            head_weights: [0.0f32; H],
             head_weights_f32: [0.0f32; H],
             head_bias: 0.0,
             use_f32_head: false,
